refactor storage class and remove storage set, get, delete info logs
