+++ lib/storage/storage.ts
@@ -1,75 +1,11 @@
-import { BadRequestError, InternalServerError } from 'errors/apps-sdk-error';
-import { RequestOptions } from 'types/fetch';
+import {BaseStorage} from 'lib/storage/base-storage';
 import { isDefined } from 'types/guards';
-import { ErrorResponse, IStorageInstance, Options, SetResponse, Token } from 'types/storage';
-import { fetchWrapper } from 'utils/fetch-wrapper';
-import { Logger } from 'utils/logger';
+import { ErrorResponse, IStorageInstance, Options, SetResponse } from 'types/storage';
 
-const logger = new Logger('Storage', { mondayInternal: true });
-
-const getStorageUrl = () => {
-  const url = process.env.STORAGE_URL || 'https://apps-storage.monday.com/app_storage_api/v2';
-  return url;
-};
-
-const getToken = (token?: Token, options: Options = {}): Token => {
-  const selectedToken = options.token || token;
-
-  if (!isDefined(selectedToken)) {
-    throw new BadRequestError('Missing token');
-  }
-
-  return selectedToken;
-};
-
-const generateCrudPath = (key: string, options: Options) => {
-  if (!isDefined(key)) {
-    throw new BadRequestError('Missing key');
-  }
-
-  const { shared } = options;
-  const storageUrl = getStorageUrl();
-  // eslint-disable-next-line @typescript-eslint/restrict-template-expressions
-  const fullPath = `${storageUrl}/${key}?shareGlobally=${!!shared}`;
-  return fullPath;
-};
-
-const storageFetch = async <T>(key: string, initToken: Token | undefined, externalOptions: Options, options: RequestOptions) => {
-  const { method, body } = options;
-  const token = getToken(initToken, externalOptions);
-  const stringifiedBody = JSON.stringify(body);
-  const url = generateCrudPath(key, externalOptions);
-  if (!isDefined(method)) {
-    throw new InternalServerError('An error occurred');
-  }
-
-  const headers = {
-    Authorization: token,
-    'Content-Type': 'application/json'
-  };
-
-  let response: T | undefined;
-  try {
-    response = await fetchWrapper<T>(url, {
-      method,
-      headers,
-      ...(body && { body: stringifiedBody })
-    });
-  } catch (error: unknown) {
-    logger.debug('[storageFetch] Unexpected error occurred while communicating with storage', { error: error as Error });
-    throw new InternalServerError('An issue occurred while accessing storage');
-  }
-
-  return response as T;
-};
-
-export class Storage implements IStorageInstance {
-  constructor(private token?: Token) {
-  }
+export class Storage extends BaseStorage implements IStorageInstance {
 
   async delete(key: string, options: Options = {}) {
-    const result = await storageFetch<ErrorResponse>(key, this.token, options, {method: 'DELETE'});
-    logger.info(`[Storage.delete] Deleted data for key from storage\nkey: ${key}`, {mondayInternal: false});
+    const result = await this.storageFetch<ErrorResponse>(key, this.token, options, {method: 'DELETE'});
     if (result?.error) {
       return { error: result.error, success: false };
     } else {
@@ -78,8 +14,7 @@ export class Storage implements IStorageInstance {
   }
 
   async get<T>(key: string, options: Options = {}) {
-    const result = await storageFetch<T>(key, this.token, options, { method: 'GET' });
-    logger.info(`[Storage.get] Got data for key from storage\nkey: ${key}`, { mondayInternal: false });
+    const result = await this.storageFetch<T>(key, this.token, options, { method: 'GET' });
     if (!isDefined(result)) {
       return { success: false, value: null };
     }
@@ -90,7 +25,7 @@ export class Storage implements IStorageInstance {
   async set(key: string, value, options: Options = {}) {
     const { previousVersion } = options;
 
-    const result = await storageFetch<SetResponse>(key, this.token, options, {
+    const result = await this.storageFetch<SetResponse>(key, this.token, options, {
       method: 'POST',
       body: {
         // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
@@ -99,7 +34,6 @@ export class Storage implements IStorageInstance {
       }
     });
 
-    logger.info(`[Storage.set] Set data for key in storage\nkey: ${key}`, { mondayInternal: false });
     const { version, error } = result;
     if (version) {
       return { version, success: true };
